pyproject.toml
[project]
name = "dspant"
version = "0.1.0"
description = "DSPANT: Digital Signal Processing for Analysis Toolkit"
readme = "README.md"
authors = [
    { name = "Jesús Peñaloza", email = "jesus.penalozaa@ufl.edu" }
]
requires-python = ">=3.12"


dependencies = [
    "colorspace>=1.0.0",
    "dask-ml>=2025.1.0",
    "dask[complete]>=2025.2.0",
    "graphviz>=0.20.3",
    "h5py>=3.13.0",
    "ipycytoscape>=1.3.3",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.10.0",
    "mkl-service>=2.4.2; sys_platform != 'darwin'",
    "neo>=0.14.1",
    "numba>=0.61.0",
    "pendulum>=3.0.0",
    "plotly-resampler>=0.10.0",
    "polars>=1.21.0",
    "pyarrow>=19.0.0",
    "pydantic>=2.10.6",
    "pywavelets>=1.8.0",
    "rich>=13.9.4",
    "scipy>=1.15.1",
    "seaborn>=0.13.2",
    "ssqueezepy>=0.6.5",
    "tdt>=0.6.7",
    "toml>=0.10.2",
    "zarr>=3.0.2",
    "zarrs>=0.1.2",
]

[project.scripts]
dspant = "dspant:main"

# Reorganize optional dependencies into logical groups
[project.optional-dependencies]

# Extension packages
emgproc = ["dspant-emgproc==0.1.0"]
neuroproc = ["dspant-neuroproc==0.1.0"]
viz = ["dspant-viz==0.1.0"]

# Documentation tools
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.6.12",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-autorefs>=0.5.0",
]

# CPU-only PyTorch
cpu = [
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
]

# CUDA-enabled PyTorch
cu126 = [
    "torch>=2.6.0; sys_platform == 'win32' or sys_platform == 'linux'",
    "torchaudio>=2.6.0; sys_platform == 'win32' or sys_platform == 'linux'",
    "torch>=2.6.0; sys_platform == 'darwin'",
    "torchaudio>=2.6.0; sys_platform == 'darwin'",
]

# All extensions combined
all = [
    "dspant-emgproc==0.1.0",
    "dspant-neuroproc==0.1.0",
    "dspant-viz==0.1.0",
    "torch>=2.6.0; sys_platform == 'win32' or sys_platform == 'linux'",
    "torchaudio>=2.6.0; sys_platform == 'win32' or sys_platform == 'linux'",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
]

dev = [
    # Development tools
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
module-name = "dspant._rs"
python-packages = ["dspant"]
python-source = "src"
manifest-path = "./rust/Cargo.toml"

# uv-specific configurations
[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu126" },
    ],
]

cache-keys = [{file = "pyproject.toml"}, {file = "./rust/Cargo.toml"}, {file = "**/*.rs"}]

[tool.uv.sources]
# CPU version
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu126", extra = "cu126", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu126", extra = "cu126", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

dspant-emgproc = [{ workspace = true }]
dspant-neuroproc = [{ workspace = true }]
dspant-viz = [{ workspace = true }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.workspace]
members = [
    "./dspant_emgproc",
    "./dspant_neuroproc",
    "./dspant_app",
    "./dspant_viz",
]

# [dependency-groups]
